misc: lis3dh_mot: Make sure ODR is always updated
Configuration changes are only possible when the sensor is enabled.
The kernel keeps internally an updated configuration of the sensor, but
this is only used when the sensor is completely turned off, which
might never happen. As consequence, some configuration changes might be
lost if users apply them when the sensor is disabled.

Make sure this never happens by applying the configuration changes
(currently REG1 only) users requested as soon as the sensor is enabled.

Change-Id: Id292024dcc12666ab5c2fcaa4789491fdeb59e2a